refactor: move query to separate module
